+++ projects/ng2-pdf-viewer/src/public-api.ts
@@ -1,0 +1,2 @@
+export * from './lib/pdf-viewer/pdf-viewer.component';
+export * from './lib/pdf-viewer/typings';