+++ public_api.ts
@@ -1,2 +1,0 @@
-export * from './src/app/pdf-viewer/pdf-viewer.component';
-export * from './src/app/pdf-viewer/typings';